Add support for official love 11.4 AppImages
This new code checks for the location of certain files instead of
simply checking for the version to retain support for custom appimages
via the `source_appimage` configuration option.
Also now spaces are stripped from the game name for everything, which
somewhat changes past behaviour, because it just leads to too many
problems, as outlined in the deranged log messages that is printed, when
spaces need to be stripped.
Additionally icons are not converted if the source file is any of png,
svg, svgz or xpm as all these are supported natively by the AppImage
format. Previously every non-png icon was converted to png.